data-pipeline/cloud_functions/igdb_ingest/requirements.txt
google-cloud-storage==2.9.0
requests==2.31.0
